01 - Explorando Banco de Dados Relacionais com Python DB API/main/pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Thiago Silva de Oliveira <thiagodevoliveira@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
